chore(heureka): adds support groups
