+++ apps/heureka/src/components/Vulnerabilities/VulnerabilitiesList/VulnerabilityDetailsPanel/VulnerabilityServicesList/VulnerabilityServicesTotalCount.tsx
@@ -1,0 +1,27 @@
+/*
+ * SPDX-FileCopyrightText: 2024 SAP SE or an SAP affiliate company and Juno contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import { use } from "react"
+import { Vulnerability } from "../../../utils"
+import { ApolloQueryResult } from "@apollo/client"
+import { GetVulnerabilitiesQuery } from "../../../../../generated/graphql"
+import { getNormalizedVulnerabilitiesResponse } from "../../../utils"
+
+type VulnerabilityServicesTotalCountProps = {
+  vulnerabilitiesPromise: Promise<ApolloQueryResult<GetVulnerabilitiesQuery>>
+  vulnerabilityName: string
+}
+
+export const VulnerabilityServicesTotalCount = ({
+  vulnerabilitiesPromise,
+  vulnerabilityName,
+}: VulnerabilityServicesTotalCountProps) => {
+  const { data } = use(vulnerabilitiesPromise)
+  const { vulnerabilities } = getNormalizedVulnerabilitiesResponse(data)
+  const vulnerabilityData = vulnerabilities.find((vuln: Vulnerability) => vuln.name === vulnerabilityName)
+  const { servicesCount } = vulnerabilityData || { servicesCount: 0 }
+
+  return servicesCount
+}
