+++ apps/heureka/src/components/Vulnerabilities/VulnerabilitiesList/VulnerabilityDetailsPanel/VulnerabilityServicesList/index.tsx
@@ -1,0 +1,84 @@
+/*
+ * SPDX-FileCopyrightText: 2024 SAP SE or an SAP affiliate company and Juno contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import React, { Suspense } from "react"
+import {
+  Stack,
+  DataGrid,
+  DataGridRow,
+  DataGridHeadCell,
+  DataGridCell,
+  Spinner,
+} from "@cloudoperators/juno-ui-components"
+import SectionContentHeading from "../../../../common/SectionContentHeading"
+import { CursorPagination } from "../../../../common/CursorPagination"
+import { ApolloQueryResult } from "@apollo/client"
+import { GetVulnerabilitiesQuery } from "../../../../../generated/graphql"
+import { getNormalizedVulnerabilityServicesResponse } from "../../../utils"
+import { VulnerabilityServicesDataRows } from "./VulnerabilityServicesDataRows"
+import { VulnerabilityServicesTotalCount } from "./VulnerabilityServicesTotalCount"
+
+type VulnerabilityServicesListProps = {
+  vulnerabilityName: string
+  vulnerabilitiesPromise: Promise<ApolloQueryResult<GetVulnerabilitiesQuery>>
+  onServiceClick?: (serviceCcrn: string) => void
+  goToPage: (after?: string | null) => void
+}
+
+export const VulnerabilityServicesList = ({
+  vulnerabilityName,
+  vulnerabilitiesPromise,
+  onServiceClick,
+  goToPage,
+}: VulnerabilityServicesListProps) => {
+  return (
+    <Suspense>
+      <SectionContentHeading>
+        Services
+        <Suspense>
+          (
+          <VulnerabilityServicesTotalCount
+            vulnerabilitiesPromise={vulnerabilitiesPromise}
+            vulnerabilityName={vulnerabilityName}
+          />
+          )
+        </Suspense>
+      </SectionContentHeading>
+
+      <div className="datagrid-hover">
+        <DataGrid columns={1}>
+          <DataGridRow>
+            <DataGridHeadCell>Service</DataGridHeadCell>
+          </DataGridRow>
+          <Suspense
+            fallback={
+              <DataGridRow>
+                <DataGridCell colSpan={1}>
+                  <Stack gap="2" alignment="center">
+                    <div>Loading</div>
+                    <Spinner variant="primary"></Spinner>
+                  </Stack>
+                </DataGridCell>
+              </DataGridRow>
+            }
+          >
+            <VulnerabilityServicesDataRows
+              vulnerabilitiesPromise={vulnerabilitiesPromise}
+              vulnerabilityName={vulnerabilityName}
+              onServiceClick={onServiceClick}
+            />
+          </Suspense>
+        </DataGrid>
+      </div>
+      <Suspense>
+        <CursorPagination
+          dataNormalizationMethod={getNormalizedVulnerabilityServicesResponse}
+          dataPromise={vulnerabilitiesPromise}
+          goToPage={goToPage}
+        />
+      </Suspense>
+    </Suspense>
+  )
+}
