+++ apps/heureka/src/components/Vulnerabilities/VulnerabilitiesList/VulnerabilityDetailsPanel/VulnerabilitySupportGroups.tsx
@@ -1,0 +1,50 @@
+/*
+ * SPDX-FileCopyrightText: 2024 SAP SE or an SAP affiliate company and Juno contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import React, { use } from "react"
+import { DataGridRow, DataGridCell, Stack, Pill, DataGridHeadCell, DataGrid } from "@cloudoperators/juno-ui-components"
+import { Vulnerability } from "../../utils"
+import { ApolloQueryResult } from "@apollo/client"
+import { GetVulnerabilitiesQuery } from "../../../../generated/graphql"
+import { getNormalizedVulnerabilitiesResponse } from "../../utils"
+
+type VulnerabilitySupportGroupsProps = {
+  vulnerabilitiesPromise: Promise<ApolloQueryResult<GetVulnerabilitiesQuery>>
+  vulnerabilityName: string
+}
+
+export const VulnerabilitySupportGroups = ({
+  vulnerabilitiesPromise,
+  vulnerabilityName,
+}: VulnerabilitySupportGroupsProps) => {
+  const { data } = use(vulnerabilitiesPromise)
+  const { vulnerabilities } = getNormalizedVulnerabilitiesResponse(data)
+  const vulnerabilityData = vulnerabilities.find((vuln: Vulnerability) => vuln.name === vulnerabilityName)
+
+  if (!vulnerabilityData) {
+    return null
+  }
+
+  return (
+    <DataGrid columns={2} gridColumnTemplate="10% auto" className="mb-6">
+      <DataGridRow>
+        <DataGridHeadCell>Support Groups</DataGridHeadCell>
+        <DataGridCell>
+          <Stack gap="1" direction="horizontal" wrap>
+            {vulnerabilityData.supportGroups?.map((group: string) => (
+              <Pill
+                key={group}
+                pillKey="support_group"
+                pillKeyLabel="support_group"
+                pillValue={group}
+                pillValueLabel={group}
+              />
+            ))}
+          </Stack>
+        </DataGridCell>
+      </DataGridRow>
+    </DataGrid>
+  )
+}
