+++ apps/heureka/src/components/Vulnerabilities/VulnerabilitiesList/VulnerabilityDetailsPanel/index.tsx
@@ -3,14 +3,12 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import React, { useCallback, useState, useMemo } from "react"
+import React, { useCallback, useState, useMemo, Suspense } from "react"
 import { useNavigate, useRouteContext, useSearch } from "@tanstack/react-router"
-import { Panel, PanelBody, Container } from "@cloudoperators/juno-ui-components"
-import { MessagesProvider, Messages } from "@cloudoperators/juno-messages-provider"
+import { Panel, PanelBody } from "@cloudoperators/juno-ui-components"
 import { VulnerabilityServicesList } from "./VulnerabilityServicesList"
+import { VulnerabilitySupportGroups } from "./VulnerabilitySupportGroups"
 import { fetchVulnerabilities } from "../../../../api/fetchVulnerabilities"
-import { ApolloQueryResult } from "@apollo/client"
-import { GetVulnerabilitiesQuery } from "../../../../generated/graphql"
 
 export const VulnerabilityPanel = () => {
   const navigate = useNavigate()
@@ -18,8 +16,29 @@ export const VulnerabilityPanel = () => {
   const { vulnerability } = useSearch({ from: "/vulnerabilities/" })
   const [currentServicesCursor, setCurrentServicesCursor] = useState<string | null | undefined>(undefined)
 
-  // Create the promise when vulnerability or cursor changes
-  const vulnerabilitiesPromise = useMemo(() => {
+  // Create a stable promise for vulnerability details (including support groups)
+  // This promise doesn't change when pagination happens
+  const vulnerabilityDetailsPromise = useMemo(() => {
+    if (vulnerability) {
+      return fetchVulnerabilities({
+        queryClient,
+        apiClient,
+        filterSettings: {
+          selectedFilters: [
+            {
+              name: "name",
+              value: vulnerability,
+            },
+          ],
+        },
+        // No afterServices parameter - this is for details only
+      })
+    }
+    return undefined
+  }, [vulnerability, queryClient, apiClient])
+
+  // Create a separate promise for services list that changes with pagination
+  const servicesListPromise = useMemo(() => {
     if (vulnerability) {
       return fetchVulnerabilities({
         queryClient,
@@ -36,7 +55,7 @@ export const VulnerabilityPanel = () => {
       })
     }
     return undefined
-  }, [vulnerability, currentServicesCursor])
+  }, [vulnerability, currentServicesCursor, queryClient, apiClient])
 
   const closeVulnerabilityPanel = useCallback(() => {
     navigate({
@@ -59,30 +78,30 @@ export const VulnerabilityPanel = () => {
   )
 
   return (
-    <MessagesProvider>
-      <Panel
-        heading={!!vulnerability ? `${vulnerability} Overview` : undefined}
-        opened={!!vulnerability}
-        onClose={closeVulnerabilityPanel}
-        size="large"
-      >
-        <PanelBody>
-          <Container py px={false}>
-            <Messages />
-          </Container>
-
-          {vulnerability && vulnerabilitiesPromise && (
-            <div className="mt-6">
-              <VulnerabilityServicesList
-                vulnerabilityName={vulnerability}
-                vulnerabilitiesPromise={vulnerabilitiesPromise}
-                onServiceClick={goToServiceDetailsPage}
-                goToPage={setCurrentServicesCursor}
-              />
-            </div>
-          )}
-        </PanelBody>
-      </Panel>
-    </MessagesProvider>
+    <Panel
+      heading={!!vulnerability ? `${vulnerability} Overview` : undefined}
+      opened={!!vulnerability}
+      onClose={closeVulnerabilityPanel}
+      size="large"
+    >
+      <PanelBody>
+        {vulnerability && vulnerabilityDetailsPromise && (
+          <Suspense>
+            <VulnerabilitySupportGroups
+              vulnerabilitiesPromise={vulnerabilityDetailsPromise}
+              vulnerabilityName={vulnerability}
+            />
+          </Suspense>
+        )}
+        {vulnerability && servicesListPromise && (
+          <VulnerabilityServicesList
+            vulnerabilityName={vulnerability}
+            vulnerabilitiesPromise={servicesListPromise}
+            onServiceClick={goToServiceDetailsPage}
+            goToPage={setCurrentServicesCursor}
+          />
+        )}
+      </PanelBody>
+    </Panel>
   )
 }
