+++ apps/heureka/src/graphql/Vulnerabilities/getVulnerabilities.graphql
@@ -33,6 +33,13 @@ query GetVulnerabilities(
             }
           }
         }
+        supportGroups{
+          edges{
+            node{
+              ccrn
+            }
+          }
+        }
       }
     }
     totalCount
